pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "amqp>=5.3.1",
    "annotated-types>=0.7.0",
    "anyio>=4.12.0",
    "async-timeout>=5.0.1",
    "attrs>=25.4.0",
    "bandit>=1.9.2",
    "bcrypt>=5.0.0",
    "beanie>=2.0.1",
    "bidict>=0.23.1",
    "billiard>=4.2.4",
    "black>=25.12.0",
    "boltons>=25.0.0",
    "boolean-py>=5.0",
    "boto3>=1.42.21",
    "botocore>=1.42.21",
    "bracex>=2.6",
    "cachecontrol>=0.14.4",
    "celery>=5.6.2",
    "certifi>=2026.1.4",
    "cffi>=2.0.0",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "click-didyoumean>=0.3.1",
    "click-option-group>=0.5.9",
    "click-plugins>=1.1.1.2",
    "click-repl>=0.3.0",
    "colorama>=0.4.6",
    "cryptography>=46.0.3",
    "cyclonedx-python-lib>=11.6.0",
    "defusedxml>=0.7.1",
    "dnspython>=2.8.0",
    "ecdsa>=0.19.1",
    "email-validator>=2.3.0",
    "exceptiongroup>=1.3.1",
    "face>=24.0.0",
    "fastapi>=0.128.0",
    "firebase-admin>=6.8.0",
    "ffmpeg-python>=0.2.0",
    "filelock>=3.20.2",
    "flake8>=7.3.0",
    "future>=1.0.0",
    "glom>=25.12.0",
    "googleapis-common-protos>=1.72.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "idna>=3.11",
    "importlib-metadata>=8.7.1",
    "iniconfig>=2.3.0",
    "isort>=7.0.0",
    "jmespath>=1.0.1",
    "jq>=1.10.0",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "kombu>=5.6.2",
    "lazy-model>=0.4.0",
    "license-expression>=30.4.4",
    "markdown-it-py>=4.0.0",
    "mccabe>=0.7.0",
    "mcp>=1.25.0",
    "mdurl>=0.1.2",
    "motor>=3.7.1",
    "msgpack>=1.1.2",
    "mypy>=1.19.1",
    "mypy-extensions>=1.1.0",
    "numpy>=2.4.0",
    "oauthlib>=3.3.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-common>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-instrumentation-requests>=0.60b1",
    "opentelemetry-proto>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-semantic-conventions>=0.60b1",
    "opentelemetry-util-http>=0.60b1",
    "packageurl-python>=0.17.6",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pathspec>=0.12.1",
    "peewee>=3.18.3",
    "pillow>=12.1.0",
    "pip-api>=0.0.34",
    "pip-audit>=2.10.0",
    "pip-requirements-parser>=32.0.1",
    "platformdirs>=4.5.1",
    "pluggy>=1.6.0",
    "prompt-toolkit>=3.0.52",
    "protobuf>=6.33.2",
    "py-serializable>=2.1.0",
    "pyasn1>=0.6.1",
    "pycodestyle>=2.14.0",
    "pycparser>=2.23",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "pyflakes>=3.4.0",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.5",
    "pyotp>=2.9.0",
    "pyparsing>=3.3.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-engineio>=4.13.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "python-socketio>=5.16.0",
    "pytokens>=0.3.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "qrcode>=8.2",
    "razorpay>=2.0.0",
    "redis>=7.1.0",
    "referencing>=0.37.0",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "rich>=14.2.0",
    "rpds-py>=0.30.0",
    "rsa>=4.9.1",
    "ruamel-yaml>=0.19.1",
    "ruamel-yaml-clib>=0.2.15",
    "ruff>=0.14.10",
    "s3transfer>=0.16.0",
    "s5cmd>=0.3.3",
    "shellingham>=1.5.4",
    "simple-websocket>=1.1.0",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sortedcontainers>=2.4.0",
    "sse-starlette>=3.1.2",
    "starlette>=0.50.0",
    "stevedore>=5.6.0",
    "stripe>=14.1.0",
    "toml>=0.10.2",
    "tomli>=2.3.0",
    "typer>=0.21.0",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.2",
    "tzdata>=2025.3",
    "tzlocal>=5.3.1",
    "urllib3>=2.6.2",
    "uvicorn>=0.40.0",
    "vine>=5.1.0",
    "watchfiles>=1.1.1",
    "wcmatch>=10.1",
    "wcwidth>=0.2.14",
    "websockets>=15.0.1",
    "wrapt>=1.17.3",
    "wsproto>=1.3.2",
    "zipp>=3.23.0",
    "socketio>=0.2.1",
]
